Fix `Layout.blades_of_grade(0)` to return the right type
Previously this returned a single float, but it should have returned a list of multivectors

This also changes `clifford.bases` to include the scalar in its returned dictionary.
This seems eminently reasonable, since the scalar is itself a base.
The only weirdness is that `locals().update(layout.blades())` will now set the variable `''` to something, but nothing can read that anyway.

Adds some type annotations while we're here.