Remove auto_version
This isn't be used for anything other than its travis install script.
Delete everything but that.

This removes the LICENSE - if this commit is merged, it means @moble (the license holder) should have indicated they fine with this in the PR comments.
A link is added back to the original `auto_version` script.