Make `conformalize` return a subclass of `Layout`.
This means that `layout.up` and `layout.einf` etc are available, which makes it easier to write code that works for both g3c and g2c.

In order to make this change, most of the interesting logic in `Cl` and `conformalize` moved to private classmethods in `Layout` and `ConformalLayout`.
In future, we could consider making these methods public, and maybe deprecating the public ones.