Allow tests to pass even if `NUMBA_DISABLE_JIT` is present.
The primary issue here is that the presence of this variable determined whether `Layout` objects could be pickled.
This also works around https://github.com/numba/numba/issues/4759.

Note that this makes some tests very slow, so many are skipped when this setting is present.
The skipping may have been overly zealous, so we might want to add back tests here and there later.