Encapsulate ordering and naming conventions in separate classes.
The end goal here is to be able to more clearly separate which bits of `Layout` are specific to the algebra, and which are specific to the choice of blade names and numbers.

This changes the signature of the `Layout` function.
It's likely no one was using it, but just in case we support the old syntax and emit a deprecation warning.