Add a new clifford.taylor_expansions module (#371)
This contains series expansions of trigonometric, exponential, and hyperbolic functions.

This also tweaks the implementation of the hitzer inverse to behave slightly better on degenerate algebras, and fixes a crash when computing an inverse in `Cl(0, 0, 0)`.

`cf.general_exp(mv)` has been deprecated in favor of `cf.taylor_expansions.exp(mv)`, or the more convenient spelling `mv.exp()`. The new function is jitted, unlike the old one.

Finally, this overloads pow in numba to allow it to be used in jitted code.